+++ lib/ci.bash
@@ -77,7 +77,7 @@ function ci_fix_version_number {
 function ci_fix_version_minor {
     local version="$1";
     if [[ "$version" =~ ^${ODOO_VERSION}\.([0-9]+)\.([0-9]+)\.([0-9]+)$ ]]; then
-        echo "${ODOO_VERSION}.${BASH_REMATCH[1]}.$(( BASH_REMATCH[2] + 1)).${BASH_REMATCH[3]}";
+        echo "${ODOO_VERSION}.${BASH_REMATCH[1]}.$(( BASH_REMATCH[2] + 1)).0";
     else
         return 1;
     fi
@@ -88,7 +88,7 @@ function ci_fix_version_minor {
 function ci_fix_version_major {
     local version="$1";
     if [[ "$version" =~ ^${ODOO_VERSION}\.([0-9]+)\.([0-9]+)\.([0-9]+)$ ]]; then
-        echo "${ODOO_VERSION}.$(( BASH_REMATCH[1] + 1)).${BASH_REMATCH[2]}.${BASH_REMATCH[3]}";
+        echo "${ODOO_VERSION}.$(( BASH_REMATCH[1] + 1)).0.0";
     else
         return 1;
     fi
@@ -622,13 +622,15 @@ function ci_do_forwardport {
         $SCRIPT_NAME ci do-forward-port --help     - print this help message
 
     Options:
-        -s|--src-branch <branch>   - [required] source branch to take changes from
-        --path <path>              - path to git repository. default current ($git_path)
-        -r|--remote <name>         - name of git remote. Default: $git_remote_name
-        --fm|--forward-migrations  - Rename all migrations for source version
-                                     to new odoo-version
+        -s|--src-branch <branch>         - [required] source branch to take changes from
+        --path <path>                    - path to git repository. default current ($git_path)
+        -r|--remote <name>               - name of git remote. Default: $git_remote_name
+        --fm|--forward-migrations        - Rename all migrations for source version
+                                           to new odoo-version. Now it is enabled by default.
+        --no-fm|--no-forward-migrations  - Do not forward-port migrations
+        --migrate-modules                - [experimental] Migrate module's code
 
-        --help                     - show this help message
+        --help                           - show this help message
     ";
 
     # Parse options
@@ -640,7 +642,8 @@ function ci_do_forwardport {
     local tmp_branch;
     local src_branch;
     local dst_branch="$ODOO_VERSION";
-    local forward_migrations;
+    local forward_migrations=1;
+    local migrate_modules=0;
     while [[ $# -gt 0 ]]
     do
         local key="$1";
@@ -659,6 +662,12 @@ function ci_do_forwardport {
             ;;
             --fm|--forward-migrations)
                 forward_migrations=1;
+            ;;
+            --no|--no-forward-migrations)
+                forward_migrations=0;
+            ;;
+            --migrate-modules)
+                migrate_modules=1;
             ;;
             -h|--help|help)
                 echo -e "$usage";
@@ -741,7 +750,7 @@ function ci_do_forwardport {
 
                 local migration_dst=${migration_dst_prefix}${migration_name};
                 if [ -d "$migration_src" ] && [ ! -d "${migration_dst}" ]; then
-                    if [ -n "$forward_migrations" ]; then
+                    if [ "$forward_migrations" -eq 1 ]; then
                         echoe -e "${LBLUEC}INFO${NC}: forwarding migration ${YELLOWC}${addon_name}${NC} (${BLUEC}${migration_name}${NC}) from ${YELLOWC}${src_branch}${NC}.${BLUEC}${migration_name}${NC} to ${YELLOWC}${dst_branch}${NC}.${BLUEC}${migration_name}${NC}!";
                         git mv "$migration_src" "${migration_dst_prefix}${migration_name}";
                     else
@@ -751,6 +760,16 @@ function ci_do_forwardport {
             done
         fi
     done
+
+    if [ "$migrate_modules" -eq 1 ]; then
+        for addon_path in "${changed_addons[@]}"; do
+            local addon_name;
+            addon_name=$(addons_get_addon_name "$addon_path");
+            echoe -e "${BLUEC}Migrating addon ${YELLOWC}${addon_name}${BLUEC} ...${NC}";
+            exec_py "${ODOO_HELPER_LIB}/pylib/ci_fw_postfixes.py" --version="$ODOO_VERSION" --path="$addon_path";
+            echoe -e "${BLUEC}Migrated addon ${YELLOWC}${addon_name}${BLUEC}: ${GREENC}OK${NC}";
+        done
+    fi
 
     # Show resulting message
     if git_is_clean "$git_path"; then